backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0
black==23.12.1
isort==5.13.2
mypy==1.8.0 